test: :test_tube: migrate tests for multi channel nonce
