+++ tests/LSP6KeyManager/LSP6KeyManager.spec.ts
@@ -569,416 +569,6 @@ describe("KeyManager", () => {
     });
   });
 
-  /** @relayer */
-  describe("> testing sequential nonces (= channel 0)", () => {
-    let channelId = 0;
-    let latestNonce;
-
-    beforeEach(async () => {
-      latestNonce = await keyManager.callStatic.getNonce(
-        externalApp.address,
-        channelId
-      );
-    });
-
-    it.each([
-      { callNb: "First", newName: "Yamen", expectedNonce: latestNonce + 1 },
-      { callNb: "Second", newName: "Nour", expectedNonce: latestNonce + 1 },
-      { callNb: "Third", newName: "Huss", expectedNonce: latestNonce + 1 },
-      { callNb: "Fourth", newName: "Moussa", expectedNonce: latestNonce + 1 },
-    ])(
-      "$callNb call > nonce should increment from $latestNonce to $expectedNonce",
-      async ({ callNb, newName, expectedNonce }) => {
-        let targetContractPayload = targetContract.interface.encodeFunctionData(
-          "setName",
-          [newName]
-        );
-        let executeRelayCallPayload =
-          universalProfile.interface.encodeFunctionData("execute", [
-            OPERATIONS.CALL,
-            targetContract.address,
-            0,
-            targetContractPayload,
-          ]);
-
-        let hash = ethers.utils.solidityKeccak256(
-          ["address", "uint256", "bytes"],
-          [keyManager.address, latestNonce, executeRelayCallPayload]
-        );
-
-        let signature = await externalApp.signMessage(
-          ethers.utils.arrayify(hash)
-        );
-
-        await keyManager.executeRelayCall(
-          keyManager.address,
-          latestNonce,
-          executeRelayCallPayload,
-          signature
-        );
-
-        let fetchedName = await targetContract.callStatic.getName();
-        let nonceAfter = await keyManager.callStatic.getNonce(
-          externalApp.address,
-          0
-        );
-
-        expect(fetchedName).toEqual(newName);
-        expect(nonceAfter).toEqBN(latestNonce.add(1)); // ensure the nonce incremented
-      }
-    );
-  });
-
-  /** @relayer */
-  describe("> testing multi-channel nonces (= channel n)", () => {
-    let nonces = [0, 1];
-
-    describe("channel 1", () => {
-      let channelId = 1;
-      let names = ["Fabian", "Yamen"];
-
-      it(`First call > nonce should increment from ${nonces[0]} to ${
-        nonces[0] + 1
-      }`, async () => {
-        let nonceBefore = await keyManager.getNonce(
-          externalApp.address,
-          channelId
-        );
-        let newName = names[0];
-
-        let targetContractPayload = targetContract.interface.encodeFunctionData(
-          "setName",
-          [newName]
-        );
-        let executeRelayCallPayload =
-          universalProfile.interface.encodeFunctionData("execute", [
-            OPERATIONS.CALL,
-            targetContract.address,
-            0,
-            targetContractPayload,
-          ]);
-
-        let hash = ethers.utils.solidityKeccak256(
-          ["address", "uint256", "bytes"],
-          [keyManager.address, nonceBefore, executeRelayCallPayload]
-        );
-
-        let signature = await externalApp.signMessage(
-          ethers.utils.arrayify(hash)
-        );
-
-        await keyManager.executeRelayCall(
-          keyManager.address,
-          nonceBefore,
-          executeRelayCallPayload,
-          signature
-        );
-
-        let fetchedName = await targetContract.callStatic.getName();
-        let nonceAfter = await keyManager.callStatic.getNonce(
-          externalApp.address,
-          channelId
-        );
-
-        expect(fetchedName).toEqual(newName);
-        expect(nonceAfter).toEqBN(nonceBefore.add(1)); // ensure the nonce incremented
-      });
-
-      it(`Second call > nonce should increment from ${nonces[1]} to ${
-        nonces[1] + 1
-      }`, async () => {
-        let nonceBefore = await keyManager.getNonce(
-          externalApp.address,
-          channelId
-        );
-        let newName = names[1];
-
-        let targetContractPayload = targetContract.interface.encodeFunctionData(
-          "setName",
-          [newName]
-        );
-        let executeRelayCallPayload =
-          universalProfile.interface.encodeFunctionData("execute", [
-            OPERATIONS.CALL,
-            targetContract.address,
-            0,
-            targetContractPayload,
-          ]);
-
-        let hash = ethers.utils.solidityKeccak256(
-          ["address", "uint256", "bytes"],
-          [keyManager.address, nonceBefore, executeRelayCallPayload]
-        );
-
-        let signature = await externalApp.signMessage(
-          ethers.utils.arrayify(hash)
-        );
-
-        await keyManager.executeRelayCall(
-          keyManager.address,
-          nonceBefore,
-          executeRelayCallPayload,
-          signature
-        );
-
-        let fetchedName = await targetContract.callStatic.getName();
-        let nonceAfter = await keyManager.callStatic.getNonce(
-          externalApp.address,
-          channelId
-        );
-
-        expect(fetchedName).toEqual(newName);
-        expect(nonceAfter).toEqBN(nonceBefore.add(1)); // ensure the nonce incremented
-      });
-    });
-
-    describe("channel 2", () => {
-      let channelId = 2;
-      let names = ["Hugo", "Reto"];
-
-      it(`First call > nonce should increment from ${nonces[0]} to ${
-        nonces[0] + 1
-      }`, async () => {
-        let nonceBefore = await keyManager.getNonce(
-          externalApp.address,
-          channelId
-        );
-        let newName = names[0];
-
-        let targetContractPayload = targetContract.interface.encodeFunctionData(
-          "setName",
-          [newName]
-        );
-        let executeRelayCallPayload =
-          universalProfile.interface.encodeFunctionData("execute", [
-            OPERATIONS.CALL,
-            targetContract.address,
-            0,
-            targetContractPayload,
-          ]);
-
-        let hash = ethers.utils.solidityKeccak256(
-          ["address", "uint256", "bytes"],
-          [keyManager.address, nonceBefore, executeRelayCallPayload]
-        );
-
-        let signature = await externalApp.signMessage(
-          ethers.utils.arrayify(hash)
-        );
-
-        await keyManager.executeRelayCall(
-          keyManager.address,
-          nonceBefore,
-          executeRelayCallPayload,
-          signature
-        );
-
-        let fetchedName = await targetContract.callStatic.getName();
-        let nonceAfter = await keyManager.callStatic.getNonce(
-          externalApp.address,
-          channelId
-        );
-
-        expect(fetchedName).toEqual(newName);
-        expect(nonceAfter).toEqBN(nonceBefore.add(1)); // ensure the nonce incremented
-      });
-
-      it(`Second call > nonce should increment from ${nonces[1]} to ${
-        nonces[1] + 1
-      }`, async () => {
-        let nonceBefore = await keyManager.getNonce(
-          externalApp.address,
-          channelId
-        );
-        let newName = names[1];
-
-        let targetContractPayload = targetContract.interface.encodeFunctionData(
-          "setName",
-          [newName]
-        );
-        let executeRelayCallPayload =
-          universalProfile.interface.encodeFunctionData("execute", [
-            OPERATIONS.CALL,
-            targetContract.address,
-            0,
-            targetContractPayload,
-          ]);
-
-        let hash = ethers.utils.solidityKeccak256(
-          ["address", "uint256", "bytes"],
-          [keyManager.address, nonceBefore, executeRelayCallPayload]
-        );
-
-        let signature = await externalApp.signMessage(
-          ethers.utils.arrayify(hash)
-        );
-
-        await keyManager.executeRelayCall(
-          keyManager.address,
-          nonceBefore,
-          executeRelayCallPayload,
-          signature
-        );
-
-        let fetchedName = await targetContract.callStatic.getName();
-        let nonceAfter = await keyManager.callStatic.getNonce(
-          externalApp.address,
-          channelId
-        );
-
-        expect(fetchedName).toEqual(newName);
-        expect(nonceAfter).toEqBN(nonceBefore.add(1)); // ensure the nonce incremented
-      });
-    });
-
-    describe("channel 3", () => {
-      let channelId = 3;
-      let names = ["Jean", "Lenny"];
-
-      it(`First call > nonce should increment from ${nonces[0]} to ${
-        nonces[0] + 1
-      }`, async () => {
-        let nonceBefore = await keyManager.getNonce(
-          externalApp.address,
-          channelId
-        );
-        let newName = names[0];
-
-        let targetContractPayload = targetContract.interface.encodeFunctionData(
-          "setName",
-          [newName]
-        );
-        let executeRelayCallPayload =
-          universalProfile.interface.encodeFunctionData("execute", [
-            OPERATIONS.CALL,
-            targetContract.address,
-            0,
-            targetContractPayload,
-          ]);
-
-        let hash = ethers.utils.solidityKeccak256(
-          ["address", "uint256", "bytes"],
-          [keyManager.address, nonceBefore, executeRelayCallPayload]
-        );
-
-        let signature = await externalApp.signMessage(
-          ethers.utils.arrayify(hash)
-        );
-
-        await keyManager.executeRelayCall(
-          keyManager.address,
-          nonceBefore,
-          executeRelayCallPayload,
-          signature
-        );
-
-        let fetchedName = await targetContract.callStatic.getName();
-        let nonceAfter = await keyManager.callStatic.getNonce(
-          externalApp.address,
-          channelId
-        );
-
-        expect(fetchedName).toEqual(newName);
-        expect(nonceAfter).toEqBN(nonceBefore.add(1)); // ensure the nonce incremented
-      });
-
-      it(`Second call > nonce should increment from ${nonces[1]} to ${
-        nonces[1] + 1
-      }`, async () => {
-        let nonceBefore = await keyManager.getNonce(
-          externalApp.address,
-          channelId
-        );
-        let newName = names[1];
-
-        let targetContractPayload = targetContract.interface.encodeFunctionData(
-          "setName",
-          [newName]
-        );
-        let executeRelayCallPayload =
-          universalProfile.interface.encodeFunctionData("execute", [
-            OPERATIONS.CALL,
-            targetContract.address,
-            0,
-            targetContractPayload,
-          ]);
-
-        let hash = ethers.utils.solidityKeccak256(
-          ["address", "uint256", "bytes"],
-          [keyManager.address, nonceBefore, executeRelayCallPayload]
-        );
-
-        let signature = await externalApp.signMessage(
-          ethers.utils.arrayify(hash)
-        );
-
-        await keyManager.executeRelayCall(
-          keyManager.address,
-          nonceBefore,
-          executeRelayCallPayload,
-          signature
-        );
-
-        let fetchedName = await targetContract.callStatic.getName();
-        let nonceAfter = await keyManager.callStatic.getNonce(
-          externalApp.address,
-          channelId
-        );
-
-        expect(fetchedName).toEqual(newName);
-        expect(nonceAfter).toEqBN(nonceBefore.add(1)); // ensure the nonce incremented
-      });
-    });
-
-    describe("channel 15", () => {
-      let channelId = 15;
-      it("First call > nonce should increment from 0 to 1", async () => {
-        let nonceBefore = await keyManager.getNonce(
-          externalApp.address,
-          channelId
-        );
-        let newName = "Lukasz";
-
-        let targetContractPayload = targetContract.interface.encodeFunctionData(
-          "setName",
-          [newName]
-        );
-        let executeRelayCallPayload =
-          universalProfile.interface.encodeFunctionData("execute", [
-            OPERATIONS.CALL,
-            targetContract.address,
-            0,
-            targetContractPayload,
-          ]);
-
-        let hash = ethers.utils.solidityKeccak256(
-          ["address", "uint256", "bytes"],
-          [keyManager.address, nonceBefore, executeRelayCallPayload]
-        );
-
-        let signature = await externalApp.signMessage(
-          ethers.utils.arrayify(hash)
-        );
-
-        await keyManager.executeRelayCall(
-          keyManager.address,
-          nonceBefore,
-          executeRelayCallPayload,
-          signature
-        );
-
-        let fetchedName = await targetContract.callStatic.getName();
-        let nonceAfter = await keyManager.callStatic.getNonce(
-          externalApp.address,
-          channelId
-        );
-
-        expect(fetchedName).toEqual(newName);
-        expect(nonceAfter).toEqBN(nonceBefore.add(1)); // ensure the nonce incremented
-      });
-    });
-  });
-
   describe("> testing Security", () => {
     let provider = ethers.provider;
     let channelId = 0;
