+++ tests/LSP6KeyManager/effects/index.ts
@@ -11,3 +11,4 @@ export * from "./AllowedAddresses.test";
 export * from "./AllowedFunctions.test";
 export * from "./AllowedStandards.test";
 export * from "./AllowedERC725YKeys.test";
+export * from "./MultiChannelNonce.test";
